factor out extinction laws
